Добавил Glide в проект, загрузку изображений в список. Вынес retrofit в отдельный модуль после ревью Владимира)
