+++ app/build.gradle
@@ -64,6 +64,10 @@ dependencies {
     implementation "org.koin:koin-android-viewmodel:2.2.1"
     implementation "org.koin:koin-android:2.2.1"
 
+    //glide
+    implementation 'com.github.bumptech.glide:glide:4.11.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.11.0'
+
     //Lifecycle
     //noinspection LifecycleAnnotationProcessorWithJava8
     kapt "androidx.lifecycle:lifecycle-compiler:2.2.0"
