+++ app/src/main/java/ru/geekbrains/bookofrecipes/presentation/MainActivity.kt
@@ -24,7 +24,6 @@ class MainActivity : AppCompatActivity() {
                 as NavHostFragment
         val navController = navHostFragment.navController
 
-        //        val navController = findNavController(R.id.nav_host_fragment)
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(
