+++ app/src/main/java/ru/geekbrains/bookofrecipes/presentation/models/RecipeModelForRecycler.kt
@@ -1,0 +1,4 @@
+package ru.geekbrains.bookofrecipes.presentation.models
+
+data class RecipeModelForRecycler(val imageUrl: String?, val title: String?, val summary: String?)
+
