--- app/src/main/java/ru/geekbrains/bookofrecipes/presentation/models/RecipeView.kt
@@ -1,4 +1,0 @@
-package ru.geekbrains.bookofrecipes.presentation.models
-
-data class RecipeView(val imageUrl: String?, val title: String?, val summary: String?)
-
