+++ app/src/main/java/ru/geekbrains/bookofrecipes/presentation/ui/recipes/RecipesViewModel.kt
@@ -4,23 +4,22 @@ import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import ru.geekbrains.bookofrecipes.data.RecipesRepository
 import ru.geekbrains.bookofrecipes.data.response.RandomRecipesResponse
 import ru.geekbrains.bookofrecipes.domain.use_cases.GetRandomRecipes
-import ru.geekbrains.bookofrecipes.presentation.models.RecipeView
+import ru.geekbrains.bookofrecipes.presentation.models.RecipeModelForRecycler
 import ru.geekbrains.bookofrecipes.service.Failure
 
 class RecipesViewModel(private val getRandomRecipes: GetRandomRecipes) : ViewModel() {
 
-    private val _recipes = MutableLiveData<List<RecipeView>>()
-    val recipes: LiveData<List<RecipeView>> = _recipes
+    private val _recipes = MutableLiveData<List<RecipeModelForRecycler>>()
+    val recipes: LiveData<List<RecipeModelForRecycler>> = _recipes
 
     fun loadRandomRecipes() =
         getRandomRecipes(10) { it.fold(::handleFailure, ::handleRandomRecipes) }
 
     private fun handleRandomRecipes(randomRecipes: RandomRecipesResponse?) {
         _recipes.value = randomRecipes?.recipes?.map { recipeInfo ->
-            RecipeView(
+            RecipeModelForRecycler(
                 recipeInfo.dishImageUrl,
                 recipeInfo.dishName,
                 recipeInfo.dishSummary
