+++ app/src/main/java/ru/geekbrains/bookofrecipes/service/RecipesApp.kt
@@ -4,13 +4,14 @@ import android.app.Application
 import org.koin.core.context.startKoin
 import ru.geekbrains.bookofrecipes.service.di.module.appModule
 import ru.geekbrains.bookofrecipes.service.di.module.repoModule
+import ru.geekbrains.bookofrecipes.service.di.module.retrofitModule
 
 class RecipesApp : Application() {
     override fun onCreate() {
         super.onCreate()
 
         startKoin {
-            modules(listOf(appModule, repoModule))
+            modules(listOf(appModule, repoModule, retrofitModule))
         }
     }
 }
