+++ app/src/main/java/ru/geekbrains/bookofrecipes/service/extensions/View.kt
@@ -1,0 +1,11 @@
+package ru.geekbrains.bookofrecipes.service.extensions
+
+import android.widget.ImageView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
+
+fun ImageView.loadFromUrl(url: String?) =
+    Glide.with(this.context.applicationContext)
+        .load(url)
+        .transition(DrawableTransitionOptions.withCrossFade())
+        .into(this)
